backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
paramiko==3.5.1
apscheduler==3.11.0
python-dotenv==1.1.0
gunicorn==23.0.0
sqlalchemy==2.0.41
alembic==1.15.2
psycopg2-binary==2.9.10
